backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
billiard==4.2.4
cachetools==6.2.4
celery==5.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
defusedxml==0.7.1
distro==1.9.0
Django==4.2.27
django-cors-headers==4.3.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filetype==1.2.0
fonttools==4.61.1
fpdf2==2.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.0
langchain-core==1.2.3
langchain-google-genai==4.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
numpy==2.4.3
opencv-python-headless==4.13.0.92
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pdf2image==1.17.0
Pillow==10.1.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyMuPDF==1.26.7
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.3.0
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
